Run production data assertions in parallel
