+++ app/jobs/ledger/assert_all_requirements_job.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+class Ledger
+  class AssertRequirementsJob < ApplicationJob
+    queue_as :low
+
+    def perform
+      Ledger::AssertCtsSyncedWithHcbCodeJob.perform_later
+      Ledger::AssertCptsSyncedWithHcbCodeJob.perform_later
+      Ledger::AssertLedgerSyncedWithHcbCodeJob.perform_later
+      Ledger::AssertNoOrphanedCtsJob.perform_later
+      Ledger::AssertNoOrphanedCptsJob.perform_later
+      Ledger::AssertCemsMatchLedgerMappingJob.perform_later
+      Ledger::AssertCpemsMatchLedgerMappingJob.perform_later
+    end
+  end
+
+end
